Separate create webpack config from gatsby-node
